// src/app/models/libro.model.ts
export interface libro {
    id?: number;  
    titulo: string;
    categoria: number;
    etiqueta: number;
    precio: number;
    ISBN: string,
    formato: string;
    paginas: number;
    color: string;
    autor: string;
    sinopsis: string;
    serie: number;
    portada: string;
    fechaPublicacion: string;
    avance: number;
}